requirements.txt
certifi==2019.9.11
chardet==3.0.4
decorator==4.0.11
Django==3.1.1.
django-sendgrid-v5==0.8.1
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
imageio==2.1.2
InstagramAPI==1.0.2
moviepy==0.2.3.2
numpy==1.18.2
Pillow==7.1.1
PyJWT==1.7.1
python-dotenv==0.14.0
python-http-client==3.3.1
python-instagram==1.3.2
pytz==2019.1
requests==2.24.0
requests-toolbelt==0.7.0
sendgrid==6.2.2
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tqdm==4.49.0
twilio==6.38.0
Unipath==1.1
urllib3==1.25.10
whitenoise
beautifulsoup4
djangorestframework==3.11.1